requirements.txt
accelerate==0.27.2
azure_storage==0.37.0
bitsandbytes==0.43.0
fairseq==0.12.2
huggingface_hub==0.21.4
HyperPyYAML==1.2.2
importlib_metadata==4.6.4
joblib==1.3.2
k2==1.24.1
kaldi_io==0.9.8
kaldilm==1.15.1
kenlm==0.2.0
matplotlib==3.8.3
numba==0.59.0
numpy==1.26.4
packaging==23.2
peft==0.9.0
pygtrie==2.5.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.31.0
sacrebleu==2.4.0
scikit_learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
streamlit==1.32.0
streamlit_chat==0.1.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.17.1+cpu
tqdm==4.66.2
transformers==4.38.2
webdataset==0.2.86
